+++ src/main/java/com/github/theredbrain/bettercombatextension/mixin/client/network/ClientPlayerEntityMixin_BetterCombatReplacementMixin.java
@@ -2,7 +2,6 @@ package com.github.theredbrain.bettercombatextension.mixin.client.network;
 
 import com.github.theredbrain.bettercombatextension.BetterCombatExtension;
 import com.github.theredbrain.bettercombatextension.client.DuckMinecraftClientMixin;
-import com.github.theredbrain.bettercombatextension.compatability.ShoulderSurfingCompat;
 import com.mojang.authlib.GameProfile;
 import net.bettercombat.BetterCombat;
 import net.bettercombat.api.MinecraftClient_BetterCombat;
@@ -15,6 +14,7 @@ import net.minecraft.client.input.Input;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.world.ClientWorld;
+import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -56,9 +56,10 @@ public abstract class ClientPlayerEntityMixin_BetterCombatReplacementMixin exten
 		boolean isWeaponSwingInProgress = ((MinecraftClient_BetterCombat) this.client).isWeaponSwingInProgress();
 		ServerConfig config = BetterCombat.config;
 		double multiplier = Math.min(Math.max((double) config.movement_speed_while_attacking, 0.0), 1.0);
-		boolean isMovementPenaltyIgnored = this.getStackInHand(((DuckMinecraftClientMixin) this.client).bettercombatextension$getCurrentAttackHand()).isIn(BetterCombatExtension.IGNORES_ATTACK_MOVEMENT_PENALTY) && isWeaponSwingInProgress;
+		ItemStack activeItemStack = this.getStackInHand(((DuckMinecraftClientMixin) this.client).bettercombatextension$getCurrentAttackHand());
+		boolean isMovementPenaltyIgnored = activeItemStack.isIn(BetterCombatExtension.IGNORES_ATTACK_MOVEMENT_PENALTY) && isWeaponSwingInProgress;
+		ClientPlayerEntity clientPlayer = (ClientPlayerEntity) (Object) this;
 		if (multiplier != 1.0 && !isMovementPenaltyIgnored) {
-			ClientPlayerEntity clientPlayer = (ClientPlayerEntity) (Object) this;
 			if (!clientPlayer.hasVehicle() || config.movement_speed_effected_while_mounting) {
 				MinecraftClient_BetterCombat client = (MinecraftClient_BetterCombat) MinecraftClient.getInstance();
 				float swingProgress = client.getSwingProgress();
@@ -78,13 +79,18 @@ public abstract class ClientPlayerEntityMixin_BetterCombatReplacementMixin exten
 					var10000.movementForward = (float) ((double) var10000.movementForward * multiplier);
 					var10000 = clientPlayer.input;
 					var10000.movementSideways = (float) ((double) var10000.movementSideways * multiplier);
-
-					if (ShoulderSurfingCompat.isShoulderSurfingCameraDecoupled() && BetterCombatExtension.serverConfig.disable_player_yaw_changes_during_attacks) {
-						var10000 = clientPlayer.input;
-						var10000.movementSideways = 0.0F;
-					}
 				}
-
+			}
+		}
+		var betterCombatExtensionServerConfig = BetterCombatExtension.serverConfig;
+		boolean isMovementLockingDisabled = activeItemStack.isIn(BetterCombatExtension.DISABLES_MOVEMENT_LOCKING_DURING_ATTACK) && isWeaponSwingInProgress;
+		if (betterCombatExtensionServerConfig.enable_movement_locking_attacks && !isMovementLockingDisabled) {
+			boolean isVehicleDisablingMovementLocking = clientPlayer.getVehicle() != null && clientPlayer.getVehicle().getType().isIn(BetterCombatExtension.DISABLES_MOVEMENT_LOCKING_WHEN_RIDDEN);
+			if (!clientPlayer.hasVehicle() || isVehicleDisablingMovementLocking) {
+				Input var10000 = clientPlayer.input;
+				var10000.movementForward = 0.0F;
+				var10000 = clientPlayer.input;
+				var10000.movementSideways = 0.0F;
 			}
 		}
 	}