+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 
   Each attack now has a "attack_stamina_cost_multiplier" instead of a "attack_stamina_cost" float field. It defaults to 1.0. There is also a new entity attribute called "generic.attack_stamina_cost" which defaults to 1.0. The actual stamina cost is the product of these two values and the "global_attack_stamina_cost_multiplier" defined in the server config. This change allows the attack stamina cost to be changed in game.
 
-- added compatibility with Shoulder Surfing Reloaded
+- reworked the "disable_player_yaw_changes_during_attacks" setting
 
-  When the "disable_player_yaw_changes_during_attacks" server config option is set to true, Shoulder Surfing's decoupled camera still functions and the player orientation can no longer be changed via the movement keys
+  renamed the option to "enable_movement_locking_attacks"
+
+  the option now enables a proper movement lock during attacks
+
+  it is recommended to use the "Shoulder Surfing Reloaded" mod and enable its "decoupled camera" setting when this setting is active
+
+  the item tag "disables_movement_locking_during_attack" and the entity type tag "disables_movement_locking_when_ridden" are used to control when the movement locking is applied
 
 # 1.1.1